evrything work je me suit arrete aux proforma

+++ src/main/java/com/soosmart/facts/dto/Article/ArticleDTO.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 public record ArticleDTO(
         UUID id,
         String libelle,
-        Double prix_unitaiare
+        Float prix_unitaiare
 ) {
 }
